# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Image to Website API",
    "description": "Production backend for turning a single product photograph into a complete, working online store with deterministic Jinja rendering, AI vision intelligence, and server-managed commerce.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/media/{object_key}": {
      "get": {
        "tags": [
          "Sites & Content"
        ],
        "summary": "Media",
        "operationId": "media_media__object_key__get",
        "parameters": [
          {
            "name": "object_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Object Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/preview/{website_id}/{filename}": {
      "get": {
        "tags": [
          "Sites & Content"
        ],
        "summary": "Preview",
        "operationId": "preview_preview__website_id___filename__get",
        "parameters": [
          {
            "name": "website_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Website Id"
            }
          },
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Filename"
            }
          },
          {
            "name": "edit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Edit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/download/{website_id}": {
      "get": {
        "tags": [
          "Sites & Content"
        ],
        "summary": "Download",
        "description": "Default: a zip holding ONE self-contained index.html plus the seller's\noriginal photo. ?single=1 returns just the .html document on its own.",
        "operationId": "download_download__website_id__get",
        "parameters": [
          {
            "name": "website_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Website Id"
            }
          },
          {
            "name": "single",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Single"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/replace-image": {
      "post": {
        "tags": [
          "Sites & Content"
        ],
        "summary": "Replace Image",
        "description": "Swap the seller's photo from inside edit mode.",
        "operationId": "replace_image_replace_image_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_replace_image_replace_image_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/{website_id}/design": {
      "get": {
        "tags": [
          "Sites & Content"
        ],
        "summary": "Current Design",
        "description": "Which design this site uses, and what it could be changed to.",
        "operationId": "current_design_api__website_id__design_get",
        "parameters": [
          {
            "name": "website_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Website Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Sites & Content"
        ],
        "summary": "Change Design",
        "description": "Re-render this site in a different design.\n\nCosts nothing and calls no model: the copy, palette and images were saved\nat generation time, so only the rendering is redone. The seller asked to\nchange the look, not to be given different words.",
        "operationId": "change_design_api__website_id__design_post",
        "parameters": [
          {
            "name": "website_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Website Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_replace_image_replace_image_post": {
        "properties": {
          "website_id": {
            "type": "string",
            "title": "Website Id"
          },
          "image": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Image"
          }
        },
        "type": "object",
        "required": [
          "website_id",
          "image"
        ],
        "title": "Body_replace_image_replace_image_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}